requirements.txt
requests==2.25.1
click==7.1.2
Pillow==8.2.0